Merged PR 2069: Handle trajectories in Store.Messages
We need to handle trajectories in Dr.Q and I've noticed that these are not currently handled in Store.Messages. Seems like quite a straightforward extension (I've copied over from Well), but I was wondering if that is going to affect any other clients of the Store.Messages API.

Additionally I've updated the Kafka libraries to their latest stable versions so that they are matching what we use in Dr.Q internally.

Related work items: #46754